# Refactor RB module for future extensions