--- a/qiskit_experiments/library/randomized_benchmarking/clifford_utils.py
+++ b/qiskit_experiments/library/randomized_benchmarking/clifford_utils.py
@@ -27,37 +27,37 @@ from qiskit.quantum_info import Clifford, random_clifford
 
 
 @lru_cache(maxsize=24)
-def clifford_1q_from_index(num):
+def clifford_1q_from_int(num):
     """Return the 1-qubit clifford element corresponding to `num`
     where `num` is between 0 and 23.
     """
-    return Clifford(clifford_1q_index_to_circuit(num), validate=False)
+    return Clifford(clifford_1q_int_to_circuit(num), validate=False)
 
 
 @lru_cache(maxsize=2048)
-def clifford_2q_from_index(num):
+def clifford_2q_from_int(num):
     """Return the 2-qubit clifford element corresponding to `num`
     where `num` is between 0 and 11519.
     """
-    return Clifford(clifford_2q_index_to_circuit(num), validate=False)
+    return Clifford(clifford_2q_int_to_circuit(num), validate=False)
 
 
-@lru_cache
-def _clifford_1q_index_to_instruction(
+@lru_cache(maxsize=None)
+def _clifford_1q_int_to_instruction(
     num: Integral, basis_gates: Optional[Tuple[str]] = None
 ) -> Instruction:
-    return clifford_1q_index_to_circuit(num, basis_gates).to_instruction()
+    return clifford_1q_int_to_circuit(num, basis_gates).to_instruction()
 
 
 @lru_cache(maxsize=2048)
-def _clifford_2q_index_to_instruction(
+def _clifford_2q_int_to_instruction(
     num: Integral, basis_gates: Optional[Tuple[str]] = None
 ) -> Instruction:
-    return clifford_2q_index_to_circuit(num, basis_gates).to_instruction()
+    return clifford_2q_int_to_circuit(num, basis_gates).to_instruction()
 
 
-@lru_cache
-def clifford_1q_index_to_circuit(
+@lru_cache(maxsize=None)
+def clifford_1q_int_to_circuit(
     num: Integral, basis_gates: Optional[Tuple[str]] = None
 ) -> QuantumCircuit:
     """Return the 1-qubit clifford circuit corresponding to `num`
@@ -94,7 +94,7 @@ _CLIFFORD_2_QUBIT_SIGS = (
 
 
 @lru_cache(maxsize=2048)
-def clifford_2q_index_to_circuit(
+def clifford_2q_int_to_circuit(
     num: Integral, basis_gates: Optional[Tuple[str]] = None
 ) -> QuantumCircuit:
     """Return the 2-qubit clifford circuit corresponding to `num`