--- a/qiskit_experiments/library/randomized_benchmarking/clifford_utils.py
+++ b/qiskit_experiments/library/randomized_benchmarking/clifford_utils.py
@@ -15,168 +15,24 @@ Utilities for using the Clifford group in randomized benchmarking
 
 from functools import lru_cache
 from numbers import Integral
-from typing import Optional, Union, Tuple
+from typing import Optional, Union
 
 from numpy.random import Generator, default_rng
 
-from qiskit import transpile
 from qiskit.circuit import Gate, Instruction
 from qiskit.circuit import QuantumCircuit, QuantumRegister
-from qiskit.circuit.library import SdgGate, HGate, SGate, SXdgGate
+from qiskit.circuit.library import SdgGate, HGate, SGate
 from qiskit.quantum_info import Clifford, random_clifford
 
 
-@lru_cache(maxsize=24)
-def clifford_1q_from_int(num):
-    """Return the 1-qubit clifford element corresponding to `num`
-    where `num` is between 0 and 23.
-    """
-    return Clifford(clifford_1q_int_to_circuit(num), validate=False)
-
-
-@lru_cache(maxsize=2048)
-def clifford_2q_from_int(num):
-    """Return the 2-qubit clifford element corresponding to `num`
-    where `num` is between 0 and 11519.
-    """
-    return Clifford(clifford_2q_int_to_circuit(num), validate=False)
-
-
 @lru_cache(maxsize=None)
-def _clifford_1q_int_to_instruction(
-    num: Integral, basis_gates: Optional[Tuple[str]] = None
-) -> Instruction:
-    return clifford_1q_int_to_circuit(num, basis_gates).to_instruction()
+def _clifford_1q_int_to_instruction(num: Integral) -> Instruction:
+    return CliffordUtils.clifford_1_qubit_circuit(num).to_instruction()
 
 
-@lru_cache(maxsize=2048)
-def _clifford_2q_int_to_instruction(
-    num: Integral, basis_gates: Optional[Tuple[str]] = None
-) -> Instruction:
-    return clifford_2q_int_to_circuit(num, basis_gates).to_instruction()
-
-
-@lru_cache(maxsize=None)
-def clifford_1q_int_to_circuit(
-    num: Integral, basis_gates: Optional[Tuple[str]] = None
-) -> QuantumCircuit:
-    """Return the 1-qubit clifford circuit corresponding to `num`
-    where `num` is between 0 and 23.
-    """
-    unpacked = _unpack_num(num, (2, 3, 4))
-    i, j, p = unpacked[0], unpacked[1], unpacked[2]
-    qc = QuantumCircuit(1, name=f"Clifford-1Q({num})")
-    if i == 1:
-        qc.h(0)
-    if j == 1:
-        qc.sxdg(0)
-    if j == 2:
-        qc.s(0)
-    if p == 1:
-        qc.x(0)
-    if p == 2:
-        qc.y(0)
-    if p == 3:
-        qc.z(0)
-
-    if basis_gates:
-        qc = transpile(qc, basis_gates=list(basis_gates))
-
-    return qc
-
-
-_CLIFFORD_2_QUBIT_SIGS = (
-    (2, 2, 3, 3, 4, 4),
-    (2, 2, 3, 3, 3, 3, 4, 4),
-    (2, 2, 3, 3, 3, 3, 4, 4),
-    (2, 2, 3, 3, 4, 4),
-)
-
-
-@lru_cache(maxsize=2048)
-def clifford_2q_int_to_circuit(
-    num: Integral, basis_gates: Optional[Tuple[str]] = None
-) -> QuantumCircuit:
-    """Return the 2-qubit clifford circuit corresponding to `num`
-    where `num` is between 0 and 11519.
-    """
-    vals = _unpack_num_multi_sigs(num, _CLIFFORD_2_QUBIT_SIGS)
-    qc = QuantumCircuit(2, name=f"Clifford-2Q({num})")
-    if vals[0] == 0 or vals[0] == 3:
-        (form, i0, i1, j0, j1, p0, p1) = vals
-    else:
-        (form, i0, i1, j0, j1, k0, k1, p0, p1) = vals
-    if i0 == 1:
-        qc.h(0)
-    if i1 == 1:
-        qc.h(1)
-    if j0 == 1:
-        qc.sxdg(0)
-    if j0 == 2:
-        qc.s(0)
-    if j1 == 1:
-        qc.sxdg(1)
-    if j1 == 2:
-        qc.s(1)
-    if form in (1, 2, 3):
-        qc.cx(0, 1)
-    if form in (2, 3):
-        qc.cx(1, 0)
-    if form == 3:
-        qc.cx(0, 1)
-    if form in (1, 2):
-        if k0 == 1:  # V gate
-            qc.sdg(0)
-            qc.h(0)
-        if k0 == 2:  # W gate
-            qc.h(0)
-            qc.s(0)
-        if k1 == 1:  # V gate
-            qc.sdg(1)
-            qc.h(1)
-        if k1 == 2:  # W gate
-            qc.h(1)
-            qc.s(1)
-    if p0 == 1:
-        qc.x(0)
-    if p0 == 2:
-        qc.y(0)
-    if p0 == 3:
-        qc.z(0)
-    if p1 == 1:
-        qc.x(1)
-    if p1 == 2:
-        qc.y(1)
-    if p1 == 3:
-        qc.z(1)
-
-    if basis_gates:
-        qc = transpile(qc, basis_gates=list(basis_gates))
-
-    return qc
-
-
-def _unpack_num(num, sig):
-    r"""Returns a tuple :math:`(a_1, \ldots, a_n)` where :math:`0 \le a_i \le \sigma_i` where
-    sig=:math:`(\sigma_1, \ldots, \sigma_n)` and num is the sequential number of the tuple
-    """
-    res = []
-    for k in sig:
-        res.append(num % k)
-        num //= k
-    return res
-
-
-def _unpack_num_multi_sigs(num, sigs):
-    """Returns the result of `_unpack_num` on one of the signatures in `sigs`"""
-    for i, sig in enumerate(sigs):
-        sig_size = 1
-        for k in sig:
-            sig_size *= k
-        if num < sig_size:
-            return [i] + _unpack_num(num, sig)
-        num -= sig_size
-    return None
+@lru_cache(maxsize=11520)
+def _clifford_2q_int_to_instruction(num: Integral) -> Instruction:
+    return CliffordUtils.clifford_2_qubit_circuit(num).to_instruction()
 
 
 class VGate(Gate):
@@ -222,17 +78,21 @@ class CliffordUtils:
         (2, 2, 3, 3, 4, 4),
     ]
 
-    def clifford_1_qubit(self, num):
+    @classmethod
+    @lru_cache(maxsize=24)
+    def clifford_1_qubit(cls, num):
         """Return the 1-qubit clifford element corresponding to `num`
         where `num` is between 0 and 23.
         """
-        return Clifford(self.clifford_1_qubit_circuit(num), validate=False)
+        return Clifford(cls.clifford_1_qubit_circuit(num), validate=False)
 
-    def clifford_2_qubit(self, num):
+    @classmethod
+    @lru_cache(maxsize=11520)
+    def clifford_2_qubit(cls, num):
         """Return the 2-qubit clifford element corresponding to `num`
         where `num` is between 0 and 11519.
         """
-        return Clifford(self.clifford_2_qubit_circuit(num), validate=False)
+        return Clifford(cls.clifford_2_qubit_circuit(num), validate=False)
 
     def random_cliffords(
         self, num_qubits: int, size: int = 1, rng: Optional[Union[int, Generator]] = None
@@ -274,38 +134,38 @@ class CliffordUtils:
             samples = rng.integers(11520, size=size)
             return [self.clifford_2_qubit_circuit(i) for i in samples]
 
+    @classmethod
     @lru_cache(maxsize=24)
-    def clifford_1_qubit_circuit(self, num):
+    def clifford_1_qubit_circuit(cls, num):
         """Return the 1-qubit clifford circuit corresponding to `num`
         where `num` is between 0 and 23.
         """
-        # pylint: disable=unbalanced-tuple-unpacking
-        # This is safe since `_unpack_num` returns list the size of the sig
-        (i, j, p) = self._unpack_num(num, self.CLIFFORD_1_QUBIT_SIG)
-        qr = QuantumRegister(1)
-        qc = QuantumCircuit(qr)
+        unpacked = cls._unpack_num(num, (2, 3, 4))
+        i, j, p = unpacked[0], unpacked[1], unpacked[2]
+        qc = QuantumCircuit(1, name=f"Clifford-1Q({num})")
         if i == 1:
             qc.h(0)
         if j == 1:
-            qc._append(SXdgGate(), [qr[0]], [])
+            qc.sxdg(0)
         if j == 2:
-            qc._append(SGate(), [qr[0]], [])
+            qc.s(0)
         if p == 1:
             qc.x(0)
         if p == 2:
             qc.y(0)
         if p == 3:
             qc.z(0)
+
         return qc
 
+    @classmethod
     @lru_cache(maxsize=11520)
-    def clifford_2_qubit_circuit(self, num):
+    def clifford_2_qubit_circuit(cls, num):
         """Return the 2-qubit clifford circuit corresponding to `num`
         where `num` is between 0 and 11519.
         """
-        vals = self._unpack_num_multi_sigs(num, self.CLIFFORD_2_QUBIT_SIGS)
-        qr = QuantumRegister(2)
-        qc = QuantumCircuit(qr)
+        vals = cls._unpack_num_multi_sigs(num, cls.CLIFFORD_2_QUBIT_SIGS)
+        qc = QuantumCircuit(2, name=f"Clifford-2Q({num})")
         if vals[0] == 0 or vals[0] == 3:
             (form, i0, i1, j0, j1, p0, p1) = vals
         else:
@@ -329,15 +189,20 @@ class CliffordUtils:
         if form == 3:
             qc.cx(0, 1)
         if form in (1, 2):
-            if k0 == 1:
-                qc._append(VGate(), [qr[0]], [])
-            if k0 == 2:
-                qc._append(WGate(), [qr[0]], [])
-            if k1 == 1:
-                qc._append(VGate(), [qr[1]], [])
-            if k1 == 2:
-                qc._append(VGate(), [qr[1]], [])
-                qc._append(VGate(), [qr[1]], [])
+            if k0 == 1:  # V gate
+                qc.sdg(0)
+                qc.h(0)
+            if k0 == 2:  # W gate
+                qc.h(0)
+                qc.s(0)
+            if k1 == 1:  # V gate
+                qc.sdg(1)
+                qc.h(1)
+            if k1 == 2:  # W gate (replacing these with qc.h(1) + qc.s(1) causes a test failure)
+                qc.sdg(1)
+                qc.h(1)
+                qc.sdg(1)
+                qc.h(1)
         if p0 == 1:
             qc.x(0)
         if p0 == 2:
@@ -350,9 +215,11 @@ class CliffordUtils:
             qc.y(1)
         if p1 == 3:
             qc.z(1)
+
         return qc
 
-    def _unpack_num(self, num, sig):
+    @staticmethod
+    def _unpack_num(num, sig):
         r"""Returns a tuple :math:`(a_1, \ldots, a_n)` where
         :math:`0 \le a_i \le \sigma_i` where
         sig=:math:`(\sigma_1, \ldots, \sigma_n)` and num is the sequential
@@ -364,7 +231,8 @@ class CliffordUtils:
             num //= k
         return res
 
-    def _unpack_num_multi_sigs(self, num, sigs):
+    @staticmethod
+    def _unpack_num_multi_sigs(num, sigs):
         """Returns the result of `_unpack_num` on one of the
         signatures in `sigs`
         """
@@ -373,6 +241,6 @@ class CliffordUtils:
             for k in sig:
                 sig_size *= k
             if num < sig_size:
-                return [i] + self._unpack_num(num, sig)
+                return [i] + CliffordUtils._unpack_num(num, sig)
             num -= sig_size
         return None