--- a/qiskit_experiments/library/randomized_benchmarking/rb_experiment.py
+++ b/qiskit_experiments/library/randomized_benchmarking/rb_experiment.py
@@ -182,7 +182,7 @@ class StandardRB(BaseExperiment, RestlessMixin):
         """Convert a RB sequence into circuit and append the inverse to the end.
 
         Returns:
-            A RB circuit.
+            A list of RB circuits.
         """
         circuits = []
         for i, seq in enumerate(sequences):