### Summary
This commit improves the readability and extensibility of RB module implementation without API change nor performance regression.

### Details and comments
- Refactoring of private functions in StandardRB and InterleavedRB for readability and preparation for the future performance improvement planned in e.g #851 and #892.