@itoko 2022-08-26
Simple benchmarking results: Suggesting no performance regression in this commit.
```
1Q StandardRB (lengths = np.arange(1, 1000, 100), num_samples = 10):
- circuits()
this commit: 1.72 s ± 57.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
main branch: 4.97 s ± 56.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- _transpile_circuits()
this commit: 16.5 s ± 163 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
main branch: 20.8 s ± 333 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

2Q StandardRB (lengths = np.arange(1, 200, 20), num_samples = 10):
- circuits()
this commit: 995 ms ± 565 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
main branch: 1.94 s ± 118 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- _transpile_circuits()
this commit: 6.84 s ± 303 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
main branch: 8.04 s ± 163 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

1Q InterleavedRB (lengths = np.arange(1, 1000, 100), num_samples = 10):
- circuits()
this commit: 5.76 s ± 174 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
main branch: 10.8 s ± 122 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- _transpile_circuits()
this commit: 54.6 s ± 830 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
main branch: 55.7 s ± 568 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

2Q InterleavedRB (lengths = np.arange(1, 200, 20), num_samples = 10):
- circuits()
this commit: 2.07 s ± 20.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
main branch: 4.11 s ± 86.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
- _transpile_circuits()
this commit: 15.3 s ± 476 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
main branch: 15.3 s ± 136 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Benchmarking code:
```
import numpy as np
from qiskit_experiments.library import StandardRB, InterleavedRB
from qiskit.circuit.library.standard_gates import SXGate
# 1Q StandardRB
rb = StandardRB(qubits=[0], lengths=np.arange(1, 1000, 100), num_samples=10, seed=1010)
# 2Q StandardRB
rb = StandardRB(qubits=[0, 1], lengths=np.arange(1, 200, 20), num_samples=10, seed=1010)
# 1Q InterleavedRB
rb = InterleavedRB(SXGate(), qubits=[0], lengths=np.arange(1, 1000, 100), num_samples=10, seed=1010)
# 2Q InterleavedRB
rb = InterleavedRB(SXGate(), qubits=[0, 1], lengths=np.arange(1, 200, 20), num_samples=10, seed=1010)
%timeit -n 1 rb.circuits()
%timeit -n 1 rb._transpiled_circuits()
```