@itoko 2022-09-05
Regarding CNOTDihedral, I think it should be implemented independently. At first I thought I could design the refactored StandardRB/InterleavedRB reusable for CNOTDihedral classes, but I had second thought. Now I think CNOTDihedral classes should not inherit StandardRB, because CNOTDihedral uses a different group to make different circuits and requires different optimizations of implementation. That said, I think the refactored StandardRB/InterleavedRB should be a good reference for implementing CNOTDihedral.