@nkanazawa1989 2022-09-05
Understood, but then I prefer to have abstract base class to restrict the flow of sequence generation for other RB variants. If some contributor writes CNOT-Dihedral RB with their research code, this would be really hard to review. I think some extra logic you introduced here can be reused for such RB variants, and we can save some boilerplate code with well-designed base class.