@itoko 2022-09-07
I think it would be better we should examine such kinds of refactoring for other RB variants later in a separate PR if necessary. (I think it would be beyond the scope of this PR)