@itoko 2022-09-08
1. I'll remove the final barrier (no one is against so far)
2. I think we still need the Clifford version for RB with 3 or more qubits even after integer Cliffords are supported for 1 and 2 qubit cases.